+++ protocol/src/main/java/org/geysermc/mcprotocollib/protocol/MinecraftProtocol.java
@@ -1,6 +1,5 @@
 package org.geysermc.mcprotocollib.protocol;
 
-import io.netty.buffer.ByteBuf;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.Setter;
@@ -10,17 +9,14 @@ import org.cloudburstmc.nbt.NbtUtils;
 import org.geysermc.mcprotocollib.auth.GameProfile;
 import org.geysermc.mcprotocollib.network.Server;
 import org.geysermc.mcprotocollib.network.Session;
-import org.geysermc.mcprotocollib.network.codec.PacketCodecHelper;
-import org.geysermc.mcprotocollib.network.codec.PacketDefinition;
 import org.geysermc.mcprotocollib.network.crypt.AESEncryption;
 import org.geysermc.mcprotocollib.network.crypt.PacketEncryption;
-import org.geysermc.mcprotocollib.network.packet.Packet;
 import org.geysermc.mcprotocollib.network.packet.PacketHeader;
 import org.geysermc.mcprotocollib.network.packet.PacketProtocol;
+import org.geysermc.mcprotocollib.network.packet.PacketRegistry;
 import org.geysermc.mcprotocollib.protocol.codec.MinecraftCodec;
 import org.geysermc.mcprotocollib.protocol.codec.MinecraftCodecHelper;
 import org.geysermc.mcprotocollib.protocol.codec.PacketCodec;
-import org.geysermc.mcprotocollib.protocol.codec.PacketStateCodec;
 import org.geysermc.mcprotocollib.protocol.data.ProtocolState;
 
 import java.io.InputStream;
@@ -49,7 +45,7 @@ public class MinecraftProtocol extends PacketProtocol {
     private final PacketCodec codec;
 
     private ProtocolState state;
-    private PacketStateCodec stateCodec;
+    private PacketRegistry stateRegistry;
 
     private final ProtocolState targetState;
 
@@ -176,6 +172,11 @@ public class MinecraftProtocol extends PacketProtocol {
         }
     }
 
+    @Override
+    public PacketRegistry getPacketRegistry() {
+        return this.stateRegistry;
+    }
+
     protected PacketEncryption enableEncryption(Key key) {
         try {
             return new AESEncryption(key);
@@ -195,57 +196,7 @@ public class MinecraftProtocol extends PacketProtocol {
 
     public void setState(ProtocolState state) {
         this.state = state;
-        this.stateCodec = this.codec.getCodec(state);
-    }
-
-    @Override
-    public Packet createClientboundPacket(int id, ByteBuf buf, PacketCodecHelper codecHelper) {
-        return this.stateCodec.createClientboundPacket(id, buf, codecHelper);
-    }
-
-    @Override
-    public int getClientboundId(Class<? extends Packet> packetClass) {
-        return this.stateCodec.getClientboundId(packetClass);
-    }
-
-    @Override
-    public int getClientboundId(Packet packet) {
-        return this.stateCodec.getClientboundId(packet);
-    }
-
-    @Override
-    public Class<? extends Packet> getClientboundClass(int id) {
-        return this.stateCodec.getClientboundClass(id);
-    }
-
-    @Override
-    public Packet createServerboundPacket(int id, ByteBuf buf, PacketCodecHelper codecHelper) {
-        return this.stateCodec.createServerboundPacket(id, buf, codecHelper);
-    }
-
-    @Override
-    public int getServerboundId(Class<? extends Packet> packetClass) {
-        return this.stateCodec.getServerboundId(packetClass);
-    }
-
-    @Override
-    public int getServerboundId(Packet packet) {
-        return this.stateCodec.getServerboundId(packet);
-    }
-
-    @Override
-    public Class<? extends Packet> getServerboundClass(int id) {
-        return this.stateCodec.getServerboundClass(id);
-    }
-
-    @Override
-    public PacketDefinition<?, ?> getServerboundDefinition(int id) {
-        return this.stateCodec.getServerboundDefinition(id);
-    }
-
-    @Override
-    public PacketDefinition<?, ?> getClientboundDefinition(int id) {
-        return this.stateCodec.getClientboundDefinition(id);
+        this.stateRegistry = this.codec.getCodec(state);
     }
 
     public static NbtMap loadNetworkCodec() {
